cbri/backend/requirements.txt
streamlit
torch
ultralytics
opencv-python
pandas
Pillow
